+++ packages/test-functions/src/models/errors.ts
@@ -5,6 +5,6 @@ export const ErrorResponseSchema = z.object({
     code: z.number().int().describe("The error code"),
     message: z.string().describe("The error message"),
 }).describe("The error response");
-app.openapiSchema('ErrorResponse', ErrorResponseSchema);
+app.openAPISchema('ErrorResponse', ErrorResponseSchema);
 
 export type ErrorResponse = z.infer<typeof ErrorResponseSchema>;